+++ com.unity.render-pipelines.universal/Runtime/ForwardRendererData.cs
@@ -25,7 +25,7 @@ namespace UnityEngine.Rendering.Universal
             }
         }
 
-        [MenuItem("Assets/Create/Rendering/Universal Render Pipeline/Forward Renderer", priority = CoreUtils.assetCreateMenuPriority2)]
+        [MenuItem("Assets/Create/Rendering/URP Forward Renderer", priority = CoreUtils.Sections.k_Section3 + CoreUtils.Priorities.k_AssetsCreateRenderingMenuPriority)]
         static void CreateForwardRendererData()
         {
             ProjectWindowUtil.StartNameEditingIfProjectWindowExists(0, CreateInstance<CreateForwardRendererAsset>(), "CustomForwardRendererData.asset", null, null);